requirements.txt
blessed==1.20.0
botocore==1.29.155
chalice==1.29.0
click==8.1.3
inquirer==2.10.1
jmespath==1.0.1
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==6.0
readchar==4.0.5
six==1.16.0
typing_extensions==4.6.3
urllib3==1.26.16
wcwidth==0.2.6
pynamodb==5.5.0
marshmallow==3.19.0
python-ulid==1.1.0
